soil_climate_agent/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
geopy
duckduckgo-search
asciichartpy
